+++ src/components/projectDetail/Zoom/HandleBox.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { ZoomContext } from '../../../context/ZoomContext';
-import * as s from '../../../style/projectDetail/ZoomimgStyle';
+import * as s from '../../../style/ZoomStyle/HandleBoxStyle';
 
 export default function HandleBox() {
   const { setShowZoomComponent, setIsZoomed, setZoomCount, updateTransform } = useContext(ZoomContext);
